requirements.txt
opencv-python==4.9.0.80
face-recognition==1.3.0
flask==3.0.2
flask-socketio==5.3.6
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1
eventlet==0.35.2
gevent==24.2.1
gevent-websocket==0.10.1
requests==2.31.0
